arquitetura de sessions
