+++ src/autentications/shared/autentication.service.ts
@@ -4,7 +4,7 @@ import { JwtService } from '@nestjs/jwt';
 import * as bcrypt from 'bcrypt';
 import { User } from '../../users/shared/user';
 import { UserService } from '../../users/services/user.service';
-import { SessionService } from '../../sessions/shared/session.service';
+import { SessionService } from '../../sessions/services/session.service';
 import { UnauthorizedError } from '../errors/unauthorized.error';
 import { UserPayload } from '../models/userPayload';
 import { UserSession } from '../models/userSession';
