+++ src/sessions/services/session.service.spec.ts
@@ -1,7 +1,8 @@
+import { getModelToken } from '@nestjs/mongoose';
 import { Test, TestingModule } from '@nestjs/testing';
-import { SessionsController } from './sessions.controller';
-import { Session } from './shared/session';
-import { SessionService } from './shared/session.service';
+import { Model } from 'mongoose';
+import { Session } from '../models/session';
+import { SessionService } from './session.service';
 
 const session: Session[] = [
   { _id: '1', user_id: '1', jwt: 'abcde1234#' },
@@ -13,40 +14,45 @@ const newSession: Session = { _id: '1', user_id: '1', jwt: 'abcde1234#' };
 
 const updatedSession = { _id: '1', user_id: '1', jwt: 'abcde1234#' };
 
-describe('SessionsController', () => {
-  let sessionController: SessionsController;
+describe('SessionsService', () => {
   let sessionService: SessionService;
+  let sessionModel: Model<Session>;
+
+  const mockUser = {
+    getAll: jest.fn().mockResolvedValue(session),
+    getById: jest.fn().mockResolvedValue(session[0]),
+    create: jest.fn().mockResolvedValue(newSession),
+    update: jest.fn().mockResolvedValue(updatedSession),
+    delete: jest.fn().mockResolvedValue(undefined),
+  };
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
-      controllers: [SessionsController],
       providers: [
         {
           provide: SessionService,
-          useValue: {
-            getAll: jest.fn().mockResolvedValue(session),
-            getById: jest.fn().mockResolvedValue(session[0]),
-            create: jest.fn().mockResolvedValue(newSession),
-            update: jest.fn().mockResolvedValue(updatedSession),
-            delete: jest.fn().mockResolvedValue(undefined),
-          },
+          useValue: mockUser,
+        },
+        {
+          provide: getModelToken('Session'),
+          useValue: mockUser,
         },
       ],
     }).compile();
 
-    sessionController = module.get<SessionsController>(SessionsController);
     sessionService = module.get<SessionService>(SessionService);
+    sessionModel = module.get<Model<Session>>(getModelToken('Session'));
   });
 
   it('should be defined', () => {
-    expect(sessionController).toBeDefined();
     expect(sessionService).toBeDefined();
+    expect(sessionModel).toBeDefined();
   });
 
   describe('getAll', () => {
     it('Deve retornar lista de usuarios logados', async () => {
       // Act
-      const result = await sessionController.getAll();
+      const result = await sessionService.getAll();
 
       // Assert
       expect(result).toEqual(session);
@@ -59,7 +65,7 @@ describe('SessionsController', () => {
       jest.spyOn(sessionService, 'getAll').mockRejectedValueOnce(new Error());
 
       // Assert
-      expect(sessionController.getAll()).rejects.toThrowError();
+      expect(sessionService.getAll()).rejects.toThrowError();
     });
   });
 
@@ -68,7 +74,7 @@ describe('SessionsController', () => {
       // Arrange
       const body: Session = { _id: '1', user_id: '1', jwt: 'abcde1234#' };
       // Act
-      const result = await sessionController.create(body);
+      const result = await sessionService.create(body);
 
       // Assert
       expect(result).toEqual(newSession);
@@ -82,13 +88,13 @@ describe('SessionsController', () => {
       jest.spyOn(sessionService, 'create').mockRejectedValueOnce(new Error());
 
       // Assert
-      expect(sessionController.create(body)).rejects.toThrowError();
+      expect(sessionService.create(body)).rejects.toThrowError();
     });
   });
   describe('getById', () => {
     it('Deve retornar uma sessão com sucesso pelo ID', async () => {
       // Act
-      const result = await sessionController.getById('1');
+      const result = await sessionService.getById('1');
 
       // Assert
       expect(result).toEqual(session[0]);
@@ -101,7 +107,7 @@ describe('SessionsController', () => {
       jest.spyOn(sessionService, 'getById').mockRejectedValueOnce(new Error());
 
       // Assert
-      expect(sessionController.getById('1')).rejects.toThrowError();
+      expect(sessionService.getById('1')).rejects.toThrowError();
     });
   });
 
@@ -111,7 +117,7 @@ describe('SessionsController', () => {
       const body: Session = { _id: '1', user_id: '1', jwt: 'abcde1234#' };
 
       // Act
-      const result = await sessionController.update('1', body);
+      const result = await sessionService.update('1', body);
 
       // Assert
       expect(result).toEqual(updatedSession);
@@ -126,7 +132,7 @@ describe('SessionsController', () => {
       jest.spyOn(sessionService, 'update').mockRejectedValueOnce(new Error());
 
       // Assert
-      expect(sessionController.update(body._id, body)).rejects.toThrowError();
+      expect(sessionService.update(body._id, body)).rejects.toThrowError();
     });
   });
 
@@ -138,7 +144,7 @@ describe('SessionsController', () => {
       };
 
       // Act
-      const result = await sessionController.delete(id._id);
+      const result = await sessionService.delete(id._id);
 
       // Assert
       expect(result).toBeUndefined();
@@ -153,7 +159,7 @@ describe('SessionsController', () => {
       jest.spyOn(sessionService, 'delete').mockRejectedValueOnce(new Error());
 
       // Assert
-      expect(sessionController.delete(id._id)).rejects.toThrowError();
+      expect(sessionService.delete(id._id)).rejects.toThrowError();
     });
   });
 });
