+++ src/sessions/services/session.service.ts
@@ -1,7 +1,7 @@
 import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
-import { Session } from './session';
+import { Session } from '../models/session';
 
 @Injectable()
 export class SessionService {
