+++ src/sessions/sessions.module.ts
@@ -1,6 +1,6 @@
 import { forwardRef, Module } from '@nestjs/common';
-import { SessionsController } from './sessions.controller';
-import { SessionService } from './shared/session.service';
+import { SessionsController } from './controllers/sessions.controller';
+import { SessionService } from './services/session.service';
 import { MongooseModule } from '@nestjs/mongoose';
 import { SessionSchema } from './schemas/session.schema';
 import { AutenticationsModule } from '../autentications/autentications.module';
